Address even more clippy lints (#805)
Satisfies and enables the following:
* `blocks_in_if_conditions`
* `comparison-chain`
* `declare_interior_mutable_const`
* `fn_to_numeric_cast`
* `comparison_chain`

+++ kernel/deferred_interrupt_tasks/src/lib.rs
@@ -68,7 +68,7 @@ pub enum InterruptRegistrationError {
     /// the interrupt handler at the given `existing_handler_address`.
     IrqInUse {
         irq: u8,
-        existing_handler_address: u64
+        existing_handler_address: usize
     },
     /// The given error occurred when spawning the deferred interrupt task.
     SpawnError(&'static str),
